refactor ingestion
