+++ Pipfile
@@ -30,3 +30,4 @@ sphinx = "==7.2.6"
 sphinx-autobuild = "==2024.2.4"
 myst-parser = "==2.0.0"
 furo = "==2024.1.29"
+pre-commit = "==4.2.0"
