+++ tests/test_file_data.py
@@ -565,18 +565,6 @@ class CQCProviderData:
 
 @dataclass
 class ONSData:
-    sample_rows = [
-        ("Yorkshire & Humber", "Leeds", "50.10101"),
-        ("Yorkshire & Humber", "York", "52.10101"),
-        ("Yorkshire & Humber", "Hull", "53.10101"),
-    ]
-
-    expected_rows = [
-        ("Yorkshire & Humber", "Leeds", "50.10101"),
-        ("Yorkshire & Humber", "York", "52.10101"),
-        ("Yorkshire & Humber", "Hull", "53.10101"),
-    ]
-
     # fmt: off
     ons_sample_rows_full = [
         ("AB10AA", "cssr1", "region1", "subicb1", "icb1", "icb_region1", "ccg1", "51.23456", "-.12345", "123", "E010123", "E020123", "Rural village", "E010123", "E020123", "pcon1", "2022", "01", "01", "20220101"),
@@ -602,73 +590,73 @@ class PAFilledPostsByIcbArea:
         ("AB10AA", date(2024,1,1), "cssr1", "icb1"),
         ("AB10AB", date(2024,1,1), "cssr1", "icb1"),
         ("AB10AC", date(2024,1,1), "cssr1", "icb1"),
-        ("AB10AA", date(2024,1,1), "cssr2", "icb2"), 
-        ("AB10AB", date(2024,1,1), "cssr2", "icb3"), 
-        ("AB10AC", date(2024,1,1), "cssr2", "icb3"), 
-        ("AB10AD", date(2024,1,1), "cssr2", "icb3"), 
+        ("AB10AA", date(2024,1,1), "cssr2", "icb2"),
+        ("AB10AB", date(2024,1,1), "cssr2", "icb3"),
+        ("AB10AC", date(2024,1,1), "cssr2", "icb3"),
+        ("AB10AD", date(2024,1,1), "cssr2", "icb3"),
         ("AB10AA", date(2023,1,1), "cssr1", "icb1"),
         ("AB10AB", date(2023,1,1), "cssr1", "icb1"),
         ("AB10AC", date(2023,1,1), "cssr1", "icb1"),
-        ("AB10AA", date(2023,1,1), "cssr2", "icb2"), 
-        ("AB10AB", date(2023,1,1), "cssr2", "icb3"), 
-        ("AB10AC", date(2023,1,1), "cssr2", "icb3"), 
+        ("AB10AA", date(2023,1,1), "cssr2", "icb2"),
+        ("AB10AB", date(2023,1,1), "cssr2", "icb3"),
+        ("AB10AC", date(2023,1,1), "cssr2", "icb3"),
     ]
 
     expected_postcode_count_per_la_rows = [
         ("AB10AA", date(2024,1,1), "cssr1", "icb1",3),
         ("AB10AB", date(2024,1,1), "cssr1", "icb1",3),
         ("AB10AC", date(2024,1,1), "cssr1", "icb1",3),
-        ("AB10AA", date(2024,1,1), "cssr2", "icb2",4), 
-        ("AB10AB", date(2024,1,1), "cssr2", "icb3",4), 
-        ("AB10AC", date(2024,1,1), "cssr2", "icb3",4), 
-        ("AB10AD", date(2024,1,1), "cssr2", "icb3",4), 
+        ("AB10AA", date(2024,1,1), "cssr2", "icb2",4),
+        ("AB10AB", date(2024,1,1), "cssr2", "icb3",4),
+        ("AB10AC", date(2024,1,1), "cssr2", "icb3",4),
+        ("AB10AD", date(2024,1,1), "cssr2", "icb3",4),
         ("AB10AA", date(2023,1,1), "cssr1", "icb1",3),
         ("AB10AB", date(2023,1,1), "cssr1", "icb1",3),
         ("AB10AC", date(2023,1,1), "cssr1", "icb1",3),
-        ("AB10AA", date(2023,1,1), "cssr2", "icb2",3), 
-        ("AB10AB", date(2023,1,1), "cssr2", "icb3",3), 
-        ("AB10AC", date(2023,1,1), "cssr2", "icb3",3), 
+        ("AB10AA", date(2023,1,1), "cssr2", "icb2",3),
+        ("AB10AB", date(2023,1,1), "cssr2", "icb3",3),
+        ("AB10AC", date(2023,1,1), "cssr2", "icb3",3),
     ]
 
     expected_postcode_count_per_la_icb_rows = [
         ("AB10AA", date(2024,1,1), "cssr1", "icb1", 3),
         ("AB10AB", date(2024,1,1), "cssr1", "icb1", 3),
         ("AB10AC", date(2024,1,1), "cssr1", "icb1", 3),
-        ("AB10AA", date(2024,1,1), "cssr2", "icb2", 1), 
-        ("AB10AB", date(2024,1,1), "cssr2", "icb3", 3), 
-        ("AB10AC", date(2024,1,1), "cssr2", "icb3", 3), 
-        ("AB10AD", date(2024,1,1), "cssr2", "icb3", 3), 
+        ("AB10AA", date(2024,1,1), "cssr2", "icb2", 1),
+        ("AB10AB", date(2024,1,1), "cssr2", "icb3", 3),
+        ("AB10AC", date(2024,1,1), "cssr2", "icb3", 3),
+        ("AB10AD", date(2024,1,1), "cssr2", "icb3", 3),
         ("AB10AA", date(2023,1,1), "cssr1", "icb1", 3),
         ("AB10AB", date(2023,1,1), "cssr1", "icb1", 3),
         ("AB10AC", date(2023,1,1), "cssr1", "icb1", 3),
-        ("AB10AA", date(2023,1,1), "cssr2", "icb2", 1), 
-        ("AB10AB", date(2023,1,1), "cssr2", "icb3", 2), 
-        ("AB10AC", date(2023,1,1), "cssr2", "icb3", 2), 
+        ("AB10AA", date(2023,1,1), "cssr2", "icb2", 1),
+        ("AB10AB", date(2023,1,1), "cssr2", "icb3", 2),
+        ("AB10AC", date(2023,1,1), "cssr2", "icb3", 2),
     ]
 
     sample_rows_with_la_and_hybrid_area_postcode_counts = [
         (date(2024,1,1), 3, 3),
-        (date(2024,1,1), 4, 1), 
-        (date(2024,1,1), 4, 3), 
+        (date(2024,1,1), 4, 1),
+        (date(2024,1,1), 4, 3),
         (date(2023,1,1), 3, 3),
-        (date(2023,1,1), 3, 1), 
+        (date(2023,1,1), 3, 1),
         (date(2023,1,1), 3, 2),
     ]
 
     expected_ratio_between_hybrid_area_and_la_area_postcodes_rows = [
         (date(2024,1,1), 3, 3, 1.00000),
-        (date(2024,1,1), 4, 1, 0.25000), 
-        (date(2024,1,1), 4, 3, 0.75000), 
+        (date(2024,1,1), 4, 1, 0.25000),
+        (date(2024,1,1), 4, 3, 0.75000),
         (date(2023,1,1), 3, 3, 1.00000),
-        (date(2023,1,1), 3, 1, 0.33333), 
+        (date(2023,1,1), 3, 1, 0.33333),
         (date(2023,1,1), 3, 2, 0.66666),
     ]
 
     full_rows_with_la_and_hybrid_area_postcode_counts = [
         ("AB10AA", date(2023,5,1), "cssr1", "icb1", 3, 3, 1.00000),
         ("AB10AB", date(2023,5,1), "cssr1", "icb1", 3, 3, 1.00000),
-        ("AB10AA", date(2023,5,1), "cssr2", "icb2", 4, 1, 0.25000), 
-        ("AB10AB", date(2023,5,1), "cssr2", "icb3", 4, 3, 0.75000), 
+        ("AB10AA", date(2023,5,1), "cssr2", "icb2", 4, 1, 0.25000),
+        ("AB10AB", date(2023,5,1), "cssr2", "icb3", 4, 3, 0.75000),
         ("AB10AA", date(2022,5,1), "cssr1", "icb1", 3, 3, 1.00000),
         ("AB10AB", date(2022,5,1), "cssr1", "icb1", 3, 3, 1.00000),
         ("AB10AC", date(2022,5,1), "cssr1", "icb1", 3, 3, 1.00000),
@@ -676,8 +664,8 @@ class PAFilledPostsByIcbArea:
 
     expected_deduplicated_import_date_hybrid_and_la_and_ratio_rows = [
         (date(2023,5,1), "cssr1", "icb1", 1.00000),
-        (date(2023,5,1), "cssr2", "icb2", 0.25000), 
-        (date(2023,5,1), "cssr2", "icb3", 0.75000), 
+        (date(2023,5,1), "cssr2", "icb2", 0.25000),
+        (date(2023,5,1), "cssr2", "icb3", 0.75000),
         (date(2022,5,1), "cssr1", "icb1", 1.00000),
     ]
     # fmt: on
@@ -712,8 +700,8 @@ class PAFilledPostsByIcbArea:
     # fmt: off
     expected_postcode_proportions_after_joining_pa_filled_posts_rows = [
         (date(2023,5,1), "Leeds", "icb1", 1.00000, 100.2, "2023"),
-        (date(2023,5,1), "Bradford", "icb2", 0.25000, 200.3, "2023"), 
-        (date(2023,5,1), "Bradford", "icb3", 0.75000, 200.3, "2023"), 
+        (date(2023,5,1), "Bradford", "icb2", 0.25000, 200.3, "2023"),
+        (date(2023,5,1), "Bradford", "icb3", 0.75000, 200.3, "2023"),
         (date(2022,5,1), "Leeds", "icb1", 1.00000, 300.3, "2022"),
         (date(2022, 5, 1), "Barking & Dagenham", "icb4", 1.00000, None, None),
     ]
@@ -3978,7 +3966,7 @@ class LmEngagementUtilsData:
         ("loc 1", date(2024, 1, 1), "cssr 1", 1, "2024"), # in ascwds on both dates
         ("loc 1", date(2024, 2, 1), "cssr 1", 1, "2024"),
         ("loc 2", date(2024, 1, 1), "cssr 2", 0, "2024"), # joins ascwds
-        ("loc 2", date(2024, 2, 1), "cssr 2", 1, "2024"), 
+        ("loc 2", date(2024, 2, 1), "cssr 2", 1, "2024"),
         ("loc 3", date(2024, 1, 1), "cssr 3", 1, "2024"), # leaves ascwds
         ("loc 3", date(2024, 2, 1), "cssr 3", 0, "2024"),
         ("loc 4", date(2024, 1, 1), "cssr 4", 0, "2024"), # multiple locations in one cssr
@@ -4129,99 +4117,94 @@ class IndCQCDataUtils:
 
     # fmt: off
     merge_map_columns_in_order_when_only_ascwds_known = [
-        ("1-001", 
+        ("1-001",
          {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-         None)
-    ]
-    #fmt: on
-
-     # fmt: off
-    expected_merge_map_columns_in_order_when_only_ascwds_known = [
-        ("1-001", 
-         {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-         None,
-         {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-         IndCQC.ascwds_job_role_ratios)
-    ]
-    #fmt: on
-
-    # fmt: off
-    merge_map_columns_in_order_when_only_primary_service_known = [
-        ("1-001", 
-         None,
-         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4})
-    ]
-    #fmt: on
-
-    # fmt: off
-    expected_merge_map_columns_in_order_when_only_primary_service_known = [
-        ("1-001", 
-         None,
-         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4},
-         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4},
-         IndCQC.ascwds_job_role_ratios_by_primary_service)
-    ]
-    #fmt: on
-
-    # fmt: off
-    merge_map_columns_in_order_when_both_map_columns_populated = [
-        ("1-001", 
-         {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4})
-    ]
-    #fmt: on
-
-    # fmt: off
-    expected_merge_map_columns_in_order_when_both_map_columns_populated = [
-        ("1-001", 
-         {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4},
-         {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-         IndCQC.ascwds_job_role_ratios)
-    ]
-    #fmt: on
-
-    # fmt: off
-    merge_map_columns_in_order_when_both_null = [
-        ("1-001", 
-         None,
-         None)
-    ]
-    #fmt: on
-
-    # fmt: off
-    expected_merge_map_columns_in_order_when_both_null = [
-        ("1-001", 
-         None,
-         None,
-         None,
          None)
     ]
     # fmt: on
 
     # fmt: off
-    merge_map_columns_in_order_when_both_map_columns_populated_at_multiple_locations = [
-        ("1-001", 
+    expected_merge_map_columns_in_order_when_only_ascwds_known = [
+        ("1-001",
          {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4}),
-        ("1-002", 
-         {MainJobRoleLabels.care_worker: 0.7, MainJobRoleLabels.registered_nurse: 0.3},
-         {MainJobRoleLabels.care_worker: 0.8, MainJobRoleLabels.registered_nurse: 0.2})
+         None,
+         {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
+         IndCQC.ascwds_job_role_ratios)
     ]
-    #fmt: on
+    # fmt: on
 
     # fmt: off
-    expected_merge_map_columns_in_order_when_both_map_columns_populated_at_multiple_locations = [
-        ("1-001", 
+    merge_map_columns_in_order_when_only_primary_service_known = [
+        ("1-001",
+         None,
+         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4})
+    ]
+    # fmt: on
+
+    # fmt: off
+    expected_merge_map_columns_in_order_when_only_primary_service_known = [
+        ("1-001",
+         None,
+         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4},
+         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4},
+         IndCQC.ascwds_job_role_ratios_by_primary_service)
+    ]
+    # fmt: on
+
+    # fmt: off
+    merge_map_columns_in_order_when_both_map_columns_populated = [
+        ("1-001",
+         {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
+         {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4})
+    ]
+    # fmt: on
+
+    # fmt: off
+    expected_merge_map_columns_in_order_when_both_map_columns_populated = [
+        ("1-001",
          {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
          {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4},
          {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
-         IndCQC.ascwds_job_role_ratios),
-        ("1-002", 
-         {MainJobRoleLabels.care_worker: 0.7, MainJobRoleLabels.registered_nurse: 0.3},
-         {MainJobRoleLabels.care_worker: 0.8, MainJobRoleLabels.registered_nurse: 0.2},
-         {MainJobRoleLabels.care_worker: 0.7, MainJobRoleLabels.registered_nurse: 0.3},
          IndCQC.ascwds_job_role_ratios)
+    ]
+    # fmt: on
+
+    merge_map_columns_in_order_when_both_null = [("1-001", None, None)]
+    expected_merge_map_columns_in_order_when_both_null = [
+        ("1-001", None, None, None, None)
+    ]
+
+    # fmt: off
+    merge_map_columns_in_order_when_both_map_columns_populated_at_multiple_locations = [
+        (
+            "1-001",
+            {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
+            {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4}
+        ),
+        (
+            "1-002",
+            {MainJobRoleLabels.care_worker: 0.7, MainJobRoleLabels.registered_nurse: 0.3},
+            {MainJobRoleLabels.care_worker: 0.8, MainJobRoleLabels.registered_nurse: 0.2}
+        )
+    ]
+    # fmt: on
+
+    # fmt: off
+    expected_merge_map_columns_in_order_when_both_map_columns_populated_at_multiple_locations = [
+        (
+            "1-001",
+            {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
+            {MainJobRoleLabels.care_worker: 0.6, MainJobRoleLabels.registered_nurse: 0.4},
+            {MainJobRoleLabels.care_worker: 0.5, MainJobRoleLabels.registered_nurse: 0.5},
+            IndCQC.ascwds_job_role_ratios
+        ),
+        (
+            "1-002",
+            {MainJobRoleLabels.care_worker: 0.7, MainJobRoleLabels.registered_nurse: 0.3},
+            {MainJobRoleLabels.care_worker: 0.8, MainJobRoleLabels.registered_nurse: 0.2},
+            {MainJobRoleLabels.care_worker: 0.7, MainJobRoleLabels.registered_nurse: 0.3},
+            IndCQC.ascwds_job_role_ratios
+        )
     ]
     # fmt: on
 
@@ -7894,7 +7877,7 @@ class ValidateLocationsAPICleanedData:
         ("1-000000002", date(2024, 1, 9), date(2024, 1, 1), "Y", "name", "prov_1", "prov_name", Sector.independent, RegistrationStatus.registered, date(2024, 1, 1), "Y", 5, PrimaryServiceType.care_home_only, date(2024, 1, 1), "cssr", "region", date(2024, 1, 1), "cssr", "region", "RUI", None, [{CQCL.name: "name", CQCL.code: "A1", CQCL.contacts: []}]),
     ]
     # fmt: on
-    
+
 
     # fmt: off
     calculate_expected_size_rows = [
@@ -9928,12 +9911,12 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsData:
 
     # fmt: off
     create_total_from_values_in_map_column_when_all_count_values_above_zero_rows = [
-        ("1-001", 
+        ("1-001",
          {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}
         )
     ]
     expected_create_total_from_values_in_map_column_when_all_count_values_above_zero_rows = [
-        ("1-001", 
+        ("1-001",
          {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2},
          3
         )
@@ -9942,12 +9925,12 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsData:
 
     # fmt: off
     create_total_from_values_in_map_column_when_all_count_values_are_null_rows = [
-        ("1-001", 
+        ("1-001",
          {MainJobRoleLabels.care_worker: None, MainJobRoleLabels.registered_nurse: None}
         )
     ]
     expected_create_total_from_values_in_map_column_when_all_count_values_are_null_rows = [
-        ("1-001", 
+        ("1-001",
          {MainJobRoleLabels.care_worker: None, MainJobRoleLabels.registered_nurse: None},
          None
         )
@@ -9977,7 +9960,7 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsData:
         expected_create_total_from_values_in_map_column_when_all_count_values_above_zero_rows
     )
     expected_create_ratios_from_counts_when_all_count_values_above_zero_rows = [
-        ("1-001", 
+        ("1-001",
          {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2},
          3,
          {MainJobRoleLabels.care_worker: 0.333, MainJobRoleLabels.registered_nurse: 0.667}),
@@ -9987,7 +9970,7 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsData:
     # fmt: off
     create_ratios_from_counts_when_all_count_values_are_null_rows = expected_create_total_from_values_in_map_column_when_all_count_values_are_null_rows
     expected_create_ratios_from_counts_when_all_count_values_are_null_rows = [
-        ("1-001", 
+        ("1-001",
          {MainJobRoleLabels.care_worker: None, MainJobRoleLabels.registered_nurse: None},
          None,
          {MainJobRoleLabels.care_worker: None, MainJobRoleLabels.registered_nurse: None},
@@ -10007,12 +9990,12 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsData:
         expected_create_total_from_values_in_map_column_at_multiple_locations_rows
     )
     expected_create_ratios_from_counts_at_multiple_establishments_rows = [
-        ("1-001", 
+        ("1-001",
          {MainJobRoleLabels.care_worker: 0, MainJobRoleLabels.registered_nurse: 1},
          1,
          {MainJobRoleLabels.care_worker: 0.0, MainJobRoleLabels.registered_nurse: 1.0}
         ),
-        ("1-002", 
+        ("1-002",
          {MainJobRoleLabels.care_worker: 2, MainJobRoleLabels.registered_nurse: 3},
          1,
          {MainJobRoleLabels.care_worker: 0.4, MainJobRoleLabels.registered_nurse: 0.6}
@@ -10064,18 +10047,18 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsData:
 
     # fmt: off
     remove_ascwds_job_role_count_when_estimate_filled_posts_source_not_ascwds_rows = [
-        ("1-001", 
-         10.0, 
+        ("1-001",
+         10.0,
          10.0,
          EstimateFilledPostsSource.ascwds_pir_merged,
          {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-        ("1-002", 
-         None, 
+        ("1-002",
+         None,
          20.0,
          EstimateFilledPostsSource.ascwds_pir_merged,
          {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-        ("1-003", 
-         10.0, 
+        ("1-003",
+         10.0,
          10.0,
          EstimateFilledPostsSource.care_home_model,
          {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
@@ -10084,18 +10067,18 @@ class EstimateIndCQCFilledPostsByJobRoleUtilsData:
 
     # fmt: off
     expected_remove_ascwds_job_role_count_when_estimate_filled_posts_source_not_ascwds_rows = [
-        ("1-001", 
-         10.0, 
+        ("1-001",
+         10.0,
          10.0,
          EstimateFilledPostsSource.ascwds_pir_merged,
          {MainJobRoleLabels.care_worker: 1, MainJobRoleLabels.registered_nurse: 2}),
-        ("1-002", 
-         None, 
+        ("1-002",
+         None,
          20.0,
          EstimateFilledPostsSource.ascwds_pir_merged,
          None),
-        ("1-003", 
-         10.0, 
+        ("1-003",
+         10.0,
          10.0,
          EstimateFilledPostsSource.care_home_model,
          None),
