+++ utils/utils.py
@@ -1,16 +1,9 @@
-import os
-import re
-import csv
 import argparse
 from typing import List, Any, Generator
 
 from pyspark.sql import DataFrame, Column, Window, SparkSession, functions as F
-from pyspark.sql.utils import AnalysisException
 
-
-import boto3
-
-TWO_MB = 2000000
+from _01_ingest.utils.utils import ingest_utils
 
 
 class SetupSpark(object):
@@ -39,54 +32,9 @@ class SetupSpark(object):
 get_spark = SetupSpark()
 
 
-def get_s3_objects_list(bucket_source, prefix, s3_resource=None):
-    if s3_resource is None:
-        s3_resource = boto3.resource("s3")
-
-    bucket_name = s3_resource.Bucket(bucket_source)
-    object_keys = []
-    for obj in bucket_name.objects.filter(Prefix=prefix):
-        if obj.size > 0:  # Ignore s3 directories
-            object_keys.append(obj.key)
-    return object_keys
-
-
 def get_model_name(path_to_model):
-    _, prefix = split_s3_uri(path_to_model)
+    _, prefix = ingest_utils.split_s3_uri(path_to_model)
     return prefix.split("/")[1]
-
-
-def read_partial_csv_content(bucket, key, s3_client=None):
-    if s3_client is None:
-        s3_client = boto3.client("s3")
-    response = s3_client.get_object(Bucket=bucket, Key=key)
-    num_bytes = int(response["ContentLength"] * 0.01)
-
-    if num_bytes > TWO_MB:
-        num_bytes = TWO_MB
-
-    return response["Body"].read(num_bytes).decode("utf-8")
-
-
-def identify_csv_delimiter(sample_csv):
-    dialect = csv.Sniffer().sniff(sample_csv, [",", "|"])
-    return dialect.delimiter
-
-
-def generate_s3_datasets_dir_date_path(
-    destination_prefix,
-    domain,
-    dataset,
-    date,
-    version="1.0.0",
-):
-    year = f"{date.year}"
-    month = f"{date.month:02d}"
-    day = f"{date.day:02d}"
-    import_date = year + month + day
-    output_dir = f"{destination_prefix}/domain={domain}/dataset={dataset}/version={version}/year={year}/month={month}/day={day}/import_date={import_date}/"
-    print(f"Generated output s3 dir: {output_dir}")
-    return output_dir
 
 
 def read_from_parquet(
@@ -119,22 +67,6 @@ def write_to_parquet(
     df.write.mode(mode).partitionBy(*partitionKeys).parquet(output_dir)
 
 
-def read_csv(source, delimiter=","):
-    spark = get_spark()
-
-    df = spark.read.option("delimiter", delimiter).csv(source, header=True)
-
-    return df
-
-
-def read_csv_with_defined_schema(source, schema):
-    spark = get_spark()
-
-    df = spark.read.schema(schema).option("header", "true").csv(source)
-
-    return df
-
-
 def format_date_fields(df, date_column_identifier="date", raw_date_format=None):
     date_columns = [column for column in df.columns if date_column_identifier in column]
 
@@ -145,34 +77,6 @@ def format_date_fields(df, date_column_identifier="date", raw_date_format=None):
             df = df.withColumn(date_column, F.to_date(date_column, raw_date_format))
 
     return df
-
-
-def is_csv(filename):
-    return filename.endswith(".csv")
-
-
-def split_s3_uri(uri):
-    bucket, prefix = uri.replace("s3://", "").split("/", 1)
-    return bucket, prefix
-
-
-def construct_s3_uri(bucket_name, key):
-    s3 = "s3://"
-    trimmed_bucket_name = bucket_name.strip()
-    s3_uri = os.path.join(s3, trimmed_bucket_name, key)
-    return s3_uri
-
-
-def get_file_directory(filepath):
-    path_delimiter = "/"
-    list_dir = filepath.split(path_delimiter)[:-1]
-    return path_delimiter.join(list_dir)
-
-
-def construct_destination_path(destination, key):
-    destination_bucket = split_s3_uri(destination)[0]
-    dir_path = get_file_directory(key)
-    return construct_s3_uri(destination_bucket, dir_path)
 
 
 def create_unix_timestamp_variable_from_date_column(
